# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/me/access": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records that the user accessed a tenant or workspace for quick access history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Record resource access",
                "parameters": [
                    {
                        "description": "Access details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.RecordAccessRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Access recorded"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/me/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current user's roles. Always includes system role if assigned.\nOptionally includes tenant role if X-Tenant-ID header is provided.\nOptionally includes workspace role if X-Workspace-ID header is provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Get my roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID to check role for",
                        "name": "X-Tenant-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID to check role for",
                        "name": "X-Workspace-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.MyRolesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/me/tenants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists tenants where the user is an active member. Supports pagination and optional search by name/code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "List my tenants with pagination and optional search",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query for tenant name or code",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.PaginatedTenantsWithRoleResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.MyRolesResponse": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.RoleEntry"
                    }
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.PaginatedTenantsWithRoleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.TenantWithRoleResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.PaginationMeta"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.RecordAccessRequest": {
            "type": "object",
            "required": [
                "entityId",
                "entityType"
            ],
            "properties": {
                "entityId": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.RoleEntry": {
            "type": "object",
            "properties": {
                "resourceId": {
                    "description": "null for SYSTEM, UUID for TENANT/WORKSPACE",
                    "type": "string"
                },
                "role": {
                    "description": "The specific role (e.g., SUPERADMIN, TENANT_OWNER, ADMIN)",
                    "type": "string"
                },
                "type": {
                    "description": "SYSTEM, TENANT, or WORKSPACE",
                    "type": "string"
                }
            }
        },
...
        "github_com_rendis_pdf-forge_core_internal_adapters_primary_http_dto.TenantWithRoleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isSystem": {
                    "type": "boolean"
                },
                "lastAccessedAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "settings": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}